Warnings(SOLO QUEDAN 16)

+++ MonkeyDelivery/Src/Control/Button.cpp
@@ -1,7 +1,7 @@
 #include "Button.h"
 #include "../Control/States/State.h"
 
-Button::Button(Game* game, int wight, int height, int x, int y) :GameObject(game), y_(0), x_(0) {
+Button::Button(Game* game, double wight, double height, int x, int y) :GameObject(game), y_(0), x_(0) {
 
 	setDimension(wight, height);
 	setPosition(x, y);
