+++ MonkeyDelivery/Src/Control/States/CreditsState.cpp
@@ -5,7 +5,7 @@
 CreditsState::CreditsState(Game* game) : State(game)
 {
 	
-	addButton(new Back1(game->getWindowWidth() / 2 - 50, game->getWindowHeight() - 250, 100, 75, game));
+	addButton(new Back1((int)game->getWindowWidth() / 2 - 50, (int)game->getWindowHeight() - 250, 100, 75, game));
 	registerCommands();
 }
