+++ MonkeyDelivery/Src/Control/UI/Sliders/BrightnessSlider.cpp
@@ -10,7 +10,7 @@ BrightnessSlider::~BrightnessSlider(){
 }
 
 void BrightnessSlider::Utility(){
-	double x =  Percent();
+	float x = (float)Percent();
 	//game_->ChangeAlphaBrightness(int(x));*/	
 	
 	SDL_SetWindowBrightness(sdlutils().window(), x);
