+++ MonkeyDelivery/Src/Logic/Shop.h
@@ -33,5 +33,5 @@ public:
 	bool buyObject(int id, int price);
 	bool inventoryFull(int selected); 
 
-	int getSize() { return objects.size(); }
+	int getSize() { return (int)objects.size(); }
 };
